[refactor] Rename the `cheat.rs` to `cheatt_sh.rs`
